chore: replace existing changes with refactored ones
